Explicitly cast integral types to double to silence a new warning in VS2019
As of version 16.7.2 Visual Studio 2019 warns about implicit int -> floating
point conversions, e.g.

C:\...\link\include\ableton/link/Beats.hpp(47,12): warning C5219:
implicit conversion from 'const int64_t' to 'double', possible loss of data

+++ include/ableton/link/Beats.hpp
@@ -44,7 +44,7 @@ struct Beats
 
   double floating() const
   {
-    return mValue / 1e6;
+    return static_cast<double>(mValue) / 1e6;
   }
 
   std::int64_t microBeats() const
